Changed the warning models all into the warning model to keep conssitency
